# add debug option and make it faster