--- a/packages/babel-plugin-core-web/lib/index.js
+++ b/packages/babel-plugin-core-web/lib/index.js
@@ -28,8 +28,14 @@ function getInjector(state) {
 			required(
 				state.opts.browsers || {
 					chrome: "41"
+				},
+				{
+					debug: state.opts.debug || false,
 				}
-			)
+			),
+			{
+				debug: state.opts.debug || false,
+			}
 		);
 		state.file.set(INJECTOR, injector);
 	}