- outputs logs in a similar style as `core-js`
- `✨  Done in 14.68s.` -> `✨  Done in 9.35s.` (main)
- `✨  Done in 56.96s.` -> `✨  Done in 19.24s.` (with updated Intl dependencies)